+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # App Center SDK for Android Change Log
 
+## Version 2.3.0
+
+### App Center Crashes
+
+* **[Feature]** Catch "low memory warning" and provide the API to check if it has happened in last session: `Crashes.hasReceivedMemoryWarningInLastSession()`.
+
+### App Center Push
+
+* **[Fix]** Fix confusing information log about the availability of the Firebase SDK.
+
+### App Center Auth
+
+* **[Feature]** App Center Auth logging now includes MSAL logs.
+* **[Fix]** Redirect URIs are now hidden in logs.
+
+### App Center Push
+
+* **[Fix]** Fix sending the push installation log after delayed start.
+
+___
+
 ## Version 2.2.0
 
 ### App Center
@@ -21,6 +42,8 @@
 * **[Feature]** Serialize `null` document values.
 * **[Fix]** Fix declaring `gson` as a build time dependency instead of runtime.
 * **[Fix]** Allow null for `ReadOptions` and `WriteOptions` parameters.
+
+___
 
 ## Version 2.1.0
 
@@ -60,6 +83,8 @@
 
 * **[Fix]** Update Firebase dependency and AppCenter push logic to avoid a runtime issue with the latest Firebase messaging version 18.0.0.
 
+___
+
 ## Version 2.0.0
 
 Version 2 of the App Center SDK includes two new modules: Auth and Data.
@@ -84,6 +109,8 @@ The App Center Data service provides functionality enabling developers to persis
 
 * **[Feature]** After calling `Auth.signIn`, the push installation is associated to the signed in user with an `accountId` and can be pushed by using the `accountId` audience. This is a different field than the `userId` set by `AppCenter.setUserId`. The push installation is also updated on calling `Auth.signOut` to stop the association.
 * **[Fix]** Fix updating push installation when setting or unsetting the user identifier by calling `AppCenter.setUserId`.
+
+___
 
 ## Version 1.11.4